[KAmillion-Base] Recording app completed
1. all record actions are now recorded by anduril and made available for playback.
all actions on the page are done through dialogue service. it is the responsibility of
dialogue to always return a promise which when resolved returns the object which
needs to plug into the playback script. before ending recording, the scrpts have to be
transposed so that all delays are relative.

Signed-off-by: himangshu hazarika <himangshu@sokratik.com>